+++ CHANGELOG.md
@@ -21,6 +21,9 @@ here: https://discord.gg/UydKWHX.
    - Sexually Inapproriate Usernames
    - Offensive Usernames
    - Banned User Names
+ * [Validation] Added wide-width unicode character username rejectors. Examples
+   of these kinds of characters can be seen
+   [here](https://www.reddit.com/r/Unicode/comments/5qa7e7/widestlongest_unicode_characters_list/).
  * [Validation] The three reaction buttons will now operate as expected:
    - White Check Mark: Verify User. User reacting needs Manage Roles.
    - Red X: Kick User. User reacting needs Kick Users.